Path generation Changes
Path generation now treats actions as blocks with set sections.

+++ Source/HonoursProject/RhythmGenerationComponent.h
@@ -49,7 +49,7 @@ public:
 	float BeatDensity;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	int BeatOccurence;
+	float BeatOccurence;
 
 	UPROPERTY(VisibleAnywhere, BlueprintReadOnly)
 	TArray<FGeneratedBeatValues> GeneratedRhythmGroup;
